Move BetaBernoulliMultiArmedBanditSurrogate to subpackage
