+++ baybe/surrogates/bandits/__init__.py
@@ -1,0 +1,1 @@
+"""Bandit surrogates."""
