+++ baybe/surrogates/bandits/beta_bernoulli.py
